feat(hack): Add simulator configuration file
This commit adds a configuration file for the skychain-sim to the hack/ directory. This file, hack/simulator.json, is configured to simulate three devices, which corresponds to the number of devices in hack/devices.json.

The simulator is configured with the following parameters:
- 3 devices
- 1-5 second interval between events
- 10% chance of a device going offline
- 5% chance of a sensor reading spike
- An endpoint of "http://localhost:8080/event"
- "temp" and "humidity" sensors

This configuration will help users to run the skychain-sim and test the SkyChain project.